Add some betting strategies
