+++ src/confseq/predmix.py
@@ -1,73 +1,5 @@
 import numpy as np
-import math
-
-
-def lambda_predmix_eb(
-    x,
-    truncation=math.inf,
-    alpha=0.05,
-    fixed_n=None,
-    prior_mean=1 / 2,
-    prior_variance=1 / 4,
-    fake_obs=1,
-    scale=1,
-):
-    """
-    Predictable mixture lambda values or "bets"
-
-    Parameters
-    ----------
-    x, array-like of 0-1 bounded reals
-        Observed data
-
-    truncation, positive real or infinity
-        Level at which to truncate lambda
-
-    alpha, (0, 1)-valued real
-        Significance level in (0, 1)
-
-    fixed_n, positive integer or None
-        Sample size for which lambda should be optimized.
-        If left as None, lambda will scale like 1/sqrt{t log t}
-
-    prior_mean, [0, 1]-valued real
-        Prior mean to use for regularized sample mean
-
-    prior_variance, (0, 1/4]-valued real
-        Prior variance to use for regularized sample variance
-
-    fake_obs, positive integer
-        Number of 'fake observations' to add.
-        Larger values correspond to more regularization near
-        `prior_mean` and `prior_variance`
-
-    scale, positive real
-        Scale by which to multiply final lambda output.
-        For most applications, this should be left as 1
-
-    Returns
-    -------
-    lambdas, array-like of positive reals
-        A (numpy) array of lambda values or "bets"
-    """
-    t = np.arange(1, len(x) + 1)
-    mu_hat_t = (fake_obs * prior_mean + np.cumsum(x)) / (t + fake_obs)
-    mu_hat_tminus1 = np.append(prior_mean, mu_hat_t[0 : (len(x) - 1)])
-    sigma2_t = (fake_obs * prior_variance + np.cumsum(np.power(x - mu_hat_t, 2))) / (
-        t + fake_obs
-    )
-    sigma2_tminus1 = np.append(prior_variance, sigma2_t[0 : (len(x) - 1)])
-    if fixed_n is None:
-        lambdas = np.sqrt(2 * np.log(1 / alpha) / (t * np.log(1 + t) * sigma2_tminus1))
-    else:
-        lambdas = np.sqrt(2 * np.log(1 / alpha) / (fixed_n * sigma2_tminus1))
-
-    lambdas[np.isnan(lambdas)] = 0
-
-    lambdas = np.minimum(truncation, lambdas)
-
-    return lambdas * scale
-
+from confseq.betting_strategies import lambda_predmix_eb
 
 def predmix_empbern_cs(
     x,
