+++ test/python_tests/test_predmix.py
@@ -1,4 +1,5 @@
 import numpy as np
+import math
 from confseq.predmix import *
